Fixed the bugs and reduced the lag
There were a couple of errors in the simulation including a very dumb one with setting the acceleration and position on the backend. Also, made it only make 3 http requests a second since too many requests means the data is not transmitted. Now it runs smoothly at 30 FPS but has a lag time between the backend and frontend of a third of a second.